requirements.txt
django>=4.2.0
firebase-admin>=6.2.0
python-decouple>=3.8
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
drf-yasg>=1.21.7 